# docs/swagger.yaml
basePath: /api
definitions:
  CreateFeedsRequest:
    properties:
      data:
        items:
          $ref: '#/definitions/CreateFeedsResource'
        type: array
    type: object
  CreateFeedsResource:
    properties:
      authors:
        example: The Owl
        type: string
        x-order: "2"
      description:
        example: Bored owls talk about whatever happens to be on their minds
        type: string
        x-order: "3"
      link:
        example: https://example.com
        type: string
        x-order: "1"
      title:
        example: Bored Owls Online Radio
        type: string
        x-order: "0"
    type: object
  CreateFeedsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/CreateFeedsResultResource'
        type: array
    type: object
  CreateFeedsResultResource:
    properties:
      id:
        example: feed_2K9BWVNuo3sG4yM322fbP3mB6ls
        type: string
        x-order: "0"
    type: object
  CreateItemResultResource:
    properties:
      id:
        example: item_2K9BWVNuo3sG4yM322fbP3mB6ls
        type: string
        x-order: "0"
    type: object
  CreateItemsRequest:
    properties:
      data:
        items:
          $ref: '#/definitions/CreateItemsResource'
        type: array
    type: object
  CreateItemsResource:
    properties:
      authors:
        example: The Owl
        type: string
        x-order: "4"
      description:
        example: Bored owls talk about whatever happens to be on their minds
        type: string
        x-order: "5"
      feed_id:
        example: feed_2K9BWVNuo3sG4yM322fbP3mB6ls
        type: string
        x-order: "1"
      file_id:
        example: file_2K9BWVNuo3sG4yM322fbP3mB6ls
        type: string
        x-order: "0"
      link:
        example: https://example.com
        type: string
        x-order: "3"
      title:
        example: Bored Owls Online Radio
        type: string
        x-order: "2"
    type: object
  CreateItemsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/CreateItemResultResource'
        type: array
    type: object
  ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  FeedResource:
    properties:
      authors:
        example: The Owl
        type: string
        x-order: "4"
      created_at:
        example: "2023-01-01T01:02:03.456Z"
        format: date-time
        type: string
        x-order: "6"
      description:
        example: Bored owls talk about whatever happens to be on their minds
        type: string
        x-order: "5"
      feed_url:
        type: string
        x-order: "1"
      id:
        example: feed_2K9BWVNuo3sG4yM322fbP3mB6ls
        type: string
        x-order: "0"
      link:
        example: https://example.com
        type: string
        x-order: "3"
      title:
        example: Bored Owls Online Radio
        type: string
        x-order: "2"
      updated_at:
        example: "2023-01-01T01:02:03.456Z"
        format: date-time
        type: string
        x-order: "7"
    type: object
  GetFeedsRequest:
    properties:
      ids:
        items:
          type: string
        type: array
        x-order: "0"
      userIds:
        items:
          type: string
        type: array
    type: object
  GetFeedsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/FeedResource'
        type: array
    type: object
  GetItemsRequest:
    properties:
      feed_ids:
        items:
          type: string
        type: array
        x-order: "1"
      ids:
        items:
          type: string
        type: array
        x-order: "0"
      userIds:
        items:
          type: string
        type: array
    type: object
  GetItemsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/ItemResource'
        type: array
    type: object
  ItemFileResource:
    properties:
      content_type:
        example: audio/mpeg
        type: string
        x-order: "3"
      id:
        example: file_2K9BWVNuo3sG4yM322fbP3mB6ls
        type: string
        x-order: "0"
      size:
        example: 123456
        type: integer
        x-order: "2"
      url:
        example: https://example.com/file.mp3
        type: string
        x-order: "1"
    type: object
  ItemResource:
    properties:
      authors:
        example: The Owl
        type: string
        x-order: "5"
      created_at:
        example: "2023-01-01T01:02:03.456Z"
        format: date-time
        type: string
        x-order: "7"
      description:
        example: Bored owls talk about whatever happens to be on their minds
        type: string
        x-order: "6"
      feed_id:
        example: feed_2K9BWVNuo3sG4yM322fbP3mB6ls
        type: string
        x-order: "2"
      file:
        allOf:
        - $ref: '#/definitions/ItemFileResource'
        x-order: "1"
      id:
        example: item_2K9BWVNuo3sG4yM322fbP3mB6ls
        type: string
        x-order: "0"
      link:
        example: https://example.com
        type: string
        x-order: "4"
      title:
        example: Bored Owls Online Radio
        type: string
        x-order: "3"
      updated_at:
        example: "2023-01-01T01:02:03.456Z"
        format: date-time
        type: string
        x-order: "8"
    type: object
  UploadFileResultResource:
    properties:
      error:
        example: invalid file format
        type: string
        x-order: "1"
      id:
        example: file_2K9BWVNuo3sG4yM322fbP3mB6ls
        type: string
        x-order: "0"
    type: object
  UploadFilesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/UploadFileResultResource'
        type: array
    type: object
info:
  contact: {}
  title: Simple Feed Service HTTP API
  version: "0.1"
paths:
  /feeds/create:
    post:
      consumes:
      - application/json
      operationId: CreateFeeds
      parameters:
      - description: CreateFeeds request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateFeedsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CreateFeedsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create new feeds
      tags:
      - Feeds
  /feeds/get:
    post:
      consumes:
      - application/json
      operationId: GetFeeds
      parameters:
      - description: Parameters for filtering feeds
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/GetFeedsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetFeedsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get feeds matching the given parameters
      tags:
      - Feeds
  /files/upload:
    post:
      consumes:
      - multipart/form-data
      operationId: UploadFiles
      parameters:
      - description: File to upload (can be repeated multiple times)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UploadFilesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Upload new audio files
      tags:
      - Files
  /items/create:
    post:
      consumes:
      - application/json
      operationId: CreateItems
      parameters:
      - description: CreateItems request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateItemsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CreateItemsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create new items and returns a response with their IDs
      tags:
      - Items
  /items/get:
    post:
      consumes:
      - application/json
      operationId: GetItems
      parameters:
      - description: Parameters for filtering items
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/GetItemsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetItemsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get items matching the given parameters
      tags:
      - Items
securityDefinitions:
  Authentication:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"